requirements.txt
djangorestframework==3.14.0
pip-chill==1.0.3
psycopg2-binary==2.9.9
python-decouple==3.8
